Fix 7, refactor edit on product detail page
